requirements.txt
discord
pynacl
python-dotenv